🚚 move http logic in http module
