+++ api/src/main.rs
@@ -1,30 +1,13 @@
-use ::domain::{
-	AggregateRootRepository, Budget, EntityRepository, Event, Payment, Project, Publisher,
-	RandomUuidGenerator, UniqueMessage, UserRepository, UuidGenerator,
-};
-use ::infrastructure::{
-	amqp::Bus,
-	config,
-	database::{self, init_pool},
-	graphql::HasuraClient,
-};
 use anyhow::Result;
 use api::{
-	domain::ProjectDetails,
 	infrastructure::database::ProjectDetailsRepository,
-	presentation::{graphql, http::routes},
+	presentation::{graphql, http},
 	Config,
 };
+use domain::{AggregateRootRepository, RandomUuidGenerator};
 use dotenv::dotenv;
-use infrastructure::tracing::Tracer;
+use infrastructure::{amqp, config, database, graphql::HasuraClient, tracing::Tracer};
 use log::info;
-use rocket::{
-	figment::{
-		providers::{Env, Format, Toml},
-		Figment,
-	},
-	routes, Build, Rocket,
-};
 use std::sync::Arc;
 use tracing::instrument;
 
@@ -35,70 +18,23 @@ async fn main() -> Result<()> {
 	let config: Config = config::load("api/app.yaml")?;
 	let _tracer = Tracer::init(config.tracer(), "api")?;
 
-	let database = Arc::new(database::Client::new(init_pool(config.database())?));
+	let database = Arc::new(database::Client::new(database::init_pool(
+		config.database(),
+	)?));
 	database.run_migrations()?;
 
-	let rocket_handler = inject_app(
-		rocket::custom(rocket_config()),
+	http::serve(
 		graphql::create_schema(),
 		Arc::new(RandomUuidGenerator),
-		Arc::new(Bus::default(config.amqp()).await?),
+		Arc::new(amqp::Bus::default(config.amqp()).await?),
 		AggregateRootRepository::new(database.clone()),
 		AggregateRootRepository::new(database.clone()),
 		AggregateRootRepository::new(database.clone()),
 		Arc::new(HasuraClient::new(config.graphql())),
 		Arc::new(ProjectDetailsRepository::new(database)),
 	)
-	.attach(routes::cors::Cors)
-	.mount(
-		"/",
-		routes![
-			routes::cors::options_preflight_handler,
-			routes::health::health_check,
-		],
-	)
-	.mount(
-		"/",
-		routes![
-			routes::graphql::graphiql,
-			routes::graphql::get_graphql_handler,
-			routes::graphql::post_graphql_handler
-		],
-	)
-	.launch();
-
-	let (rocket_result,) = tokio::join!(rocket_handler);
-	let _ = rocket_result?;
+	.await?;
 
 	info!("👋 Gracefully shut down");
 	Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-fn inject_app(
-	rocket: Rocket<Build>,
-	schema: graphql::Schema,
-	uuid_generator: Arc<dyn UuidGenerator>,
-	event_publisher: Arc<dyn Publisher<UniqueMessage<Event>>>,
-	project_repository: AggregateRootRepository<Project>,
-	payment_repository: AggregateRootRepository<Payment>,
-	budget_repository: AggregateRootRepository<Budget>,
-	user_repository: Arc<dyn UserRepository>,
-	project_details_repository: Arc<dyn EntityRepository<ProjectDetails>>,
-) -> Rocket<Build> {
-	rocket
-		.manage(schema)
-		.manage(uuid_generator)
-		.manage(event_publisher)
-		.manage(project_repository)
-		.manage(payment_repository)
-		.manage(budget_repository)
-		.manage(user_repository)
-		.manage(project_details_repository)
-}
-
-fn rocket_config() -> Figment {
-	Figment::from(rocket::Config::default())
-		.merge(Toml::file("api/Rocket.toml").nested())
-		.merge(Env::prefixed("ROCKET_").global())
 }
